Change SystemParentData::aWindow to sal_uIntPtr
...with similar reasoning as in 5d422c0348efd1df6b3d44dd6efcf5c5490fc321 "Revert
'sal_uIntPtr->unsigned long in SystemEnvData'":  Prior to
3aef606f2758172a27718a06fea0ff9080e4d80f "use tools::Long in vcl" this was long,
but besides X11 window resource IDs (for whose appropriate type see the commit
message of 5d422c0348efd1df6b3d44dd6efcf5c5490fc321), at least
vcl/unx/gtk3/gtk3gtkframe.cxx uses it in combination with GdkNativeWindow (e.g.,
GtkSalFrame::m_aForeignParentWindow and GtkSalFrame::findTopLevelSystemWindow),
which itself is a typedef for sal_uIntPtr (vcl/inc/unx/gtk/gtkframe.hxx).

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/104695
Tested-by: Jenkins

+++ toolkit/source/awt/vclxwindow1.cxx
@@ -84,7 +84,7 @@ void VCLXWindow::SetSystemParent_Impl( const css::uno::Any& rHandle )
 #elif defined( IOS )
     // Nothing
 #elif defined( UNX )
-    aSysParentData.aWindow = static_cast<long>(nHandle);
+    aSysParentData.aWindow = nHandle;
     aSysParentData.bXEmbedSupport = bXEmbed;
 #endif
